+++ internal/models/user.go
@@ -7,8 +7,8 @@ type User struct {
 	Name         string    `json:"name"`
 	Age          uint8     `json:"age"`
 	Username     string    `json:"username"`
-	Balance      float64   `json:"balance"`
-	Withdrawn    float64   `json:"withdrawn"`
+	Balance      float32   `json:"balance"`
+	Withdrawn    float32   `json:"withdrawn"`
 	Password     string    `json:"-"`
 	RefreshToken string    `json:"-"`
 }
